Remove restriction about the number of procs

+++ Tutorial/gray-scott/simulation/settings.h
@@ -6,7 +6,7 @@
 class Settings
 {
 public:
-    int L;
+    size_t L;
     int steps;
     int plotgap;
     double F;
